fixed plane hit issue
there was a problem with the way ray hits were calculated for planes: if the plane's normal was pointing in relatively the same direction as the ray (within 90 degrees of the ray) then it would return a negative distance and register a hit, even though it shouldn't have hit it. By now checking the distance returned by the intersection method, this no longer occurs.