query tree hash table
What i didn't realize at the time was that in my current set up there is a possibility that the ray intersection query for meshes may end up checking the same triangles multiple times (due to the fact that triangle pointers are duplicated in the query tree). To fix this, i've added a hash table of mesh triangle pointers to check and make sure that the current triangle has not already been evaluated.